Add files via upload
Three month Thesis Project to investigate an association of Kraft Paper in municipal waste with the growth of E-commerce. This project entailed using New York Citywide waste as an open source dataset to see an increase in the component of Kraft Paper as a constituent to make the association with the growth of E-commerce.